feat: reelShadowを外部ファイルに切り出し
