+++ src/components/Table/TableReel.tsx
@@ -1,6 +1,7 @@
 import React, { HTMLAttributes, PropsWithChildren, useRef, useState } from 'react'
 import styled, { css } from 'styled-components'
 
+import { reelShadow } from './reelShadow'
 import { useReelCells } from './useReelCells'
 
 type ElementProps = Omit<HTMLAttributes<HTMLDivElement>, keyof PropsWithChildren>
@@ -18,33 +19,6 @@ export const TableReel: React.FC<PropsWithChildren & ElementProps> = ({ children
       </Wrapper>
     </Shadow>
   )
-}
-
-type ReelShadow = {
-  showShadow?: boolean
-  direction?: 'left' | 'right'
-}
-
-export const reelShadow = ({ showShadow = true, direction = 'left' }: ReelShadow) => {
-  const shadowWidth = '12px' //影の横幅はpx指定で固定とする
-
-  return `
-    &::after {
-      content: '';
-      position: absolute;
-      z-index: 0;
-      left: ${direction === 'left' ? '0' : `-${shadowWidth}`};
-      top: 0;
-      width: ${shadowWidth};
-      pointer-events: none; /* 影の領域が広すぎるとクリッカブルエリアを侵食するので無効化 */
-      inset-block: 0;
-      background: linear-gradient(${
-        direction === 'left' ? '90deg' : '-90deg'
-      }, rgba(0, 0, 0, 0.2) 0%, rgba(0, 0, 0, 0) 100%);
-      opacity: ${showShadow ? 1 : 0};
-      transition: opacity 0.2s;
-    }
-  `
 }
 
 const Shadow = styled.div<{ showShadow: boolean }>`
