+++ src/components/Table/Td.tsx
@@ -3,7 +3,7 @@ import styled, { css } from 'styled-components'
 
 import { Theme, useTheme } from '../../hooks/useTheme'
 
-import { reelShadow } from './TableReel'
+import { reelShadow } from './reelShadow'
 import { useTdClassNames } from './useClassNames'
 
 export type Props = PropsWithChildren<{
